requirements.txt
langchain-core
langchain-community
langgraph
langchain-groq
python-dotenv
sqlalchemy
ipython
groq
mermaid
typing-extensions